rename frwc strategy
