+++ src/strategies/erc721-multi-registry-weighted/README.md
