+++ src/strategies/erc721-multi-registry-weighted/index.ts
